backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
GitPython==3.1.40
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
markdown==3.5.1
jinja2==3.1.2
python-dotenv==1.0.0 